// src/types/enums.ts
export enum FieldType {
  String = "string",
  Integer = "integer",
  Boolean = "boolean",
  Dropdown = "dropdown",
}

export enum Frequency {
  Daily = "Daily",
  Mondays = "Mondays",
  Tuesdays = "Tuesdays",
  Wednesdays = "Wednesdays",
  Thursdays = "Thursdays",
  Fridays = "Fridays",
  Saturdays = "Saturdays",
  Sundays = "Sundays",
}